Mock related functionality moved out from BasePage to a mocks util for the Card related pages & tests
