+++ packages/e2e/tests/_common/checkMocking.test.js
@@ -1,4 +1,5 @@
-import CardComponentPage from './CardComponent.page';
+import { checkSDKMocking } from './cardMocks';
+import CardComponentPage from '../_models/CardComponent.page';
 
 const cardPage = new CardComponentPage();
 
@@ -9,5 +10,5 @@ fixture`Test that bin mocking isn't turned on in the SDK`.page(cardPage.pageUrl)
  * - this is used for testing (in triggerBinLookup.ts) but if left on will break or skew the tests
  */
 test('Check for SDK Bin mocking', async () => {
-    await cardPage.checkSDKMocking();
+    await checkSDKMocking();
 });
